// eparser-cli/src/app/model/parsing-page.model.ts
export class ParsingPageModel {
  id: number;
  pageName: string;
  pageType: string;
  linkId: number;
  parseUrl: string;
  prefixUrl: string;
  pageAttr: string;
  pageCount: number;
  processed: string;
  lastUpdate: string;
  updateResult: string;
  description: string;
  tagLessInfo: string;
  tagSectionStart: string;

}